Revert "merge utils_test.go and util_test.go"
This reverts commit 1e6e8fd5016f152377e20793c244a05bf5c7831a.
